Fix issue with timer component and improve code
