+++ app/src/app/sections/lia/lia-description/lia-description.component.html
@@ -6,7 +6,7 @@
       <lia-info-time>
         <i class="material-icons">schedule</i>
         <app-lia-timer
-          [remainingSeconds]="remainingSeconds"
+          [lia]="lia"
         ></app-lia-timer>
       </lia-info-time>
     </lia-info-block>
