+++ app/src/app/sections/lia/lia-description/lia-description.component.ts
@@ -18,5 +18,4 @@ import { LiaInfoBlockComponent } from './../../../components/lia-info-block';
 
 export class LiaDescriptionComponent {
   @Input() lia: ILia;
-  @Input() remainingSeconds: Number;
 }
