+++ app/src/app/sections/lia/lia-submission-page/lia-submission-page.component.html
@@ -6,7 +6,6 @@
   ></lia-toast>
   <lia-description
     [lia]="lia"
-    [remainingSeconds]="getRemainingSeconds(lia)"
   ></lia-description>
   <lia-code-submission
     [lia]="lia"
